UC-4193 StateQueueClient 100ms hard coded timeout  - BlockingTcpClient.cpp merged back to BlockingTcpClient.h to avoid compiltaion issues
